requirements.txt
transformers==4.30.2
datasets==2.3.2
faiss-cpu==1.7.2
torch~=1.13.1
grpcio==1.62.2